+++ contrib/opencensus-ext-azure/tests/test_transport_mixin.py
@@ -27,9 +27,11 @@ from opencensus.ext.azure.common.storage import LocalFileStorage
 from opencensus.ext.azure.common.transport import (
     _MAX_CONSECUTIVE_REDIRECTS,
     _MONITOR_OAUTH_SCOPE,
+    _REACHED_INGESTION_STATUS_CODES,
     TransportMixin,
     _requests_map,
 )
+from opencensus.ext.azure.statsbeat import state
 
 TEST_FOLDER = os.path.abspath('.test.transport')
 
@@ -57,17 +59,120 @@ class MockResponse(object):
 
 # pylint: disable=W0212
 class TestTransportMixin(unittest.TestCase):
+    def setUp(self):
+        # pylint: disable=protected-access
+        _requests_map.clear()
+        state._STATSBEAT_STATE = {
+            "INITIAL_FAILURE_COUNT": 0,
+            "INITIAL_SUCCESS": False,
+            "SHUTDOWN": False,
+        }
+
     def test_check_stats_collection(self):
         mixin = TransportMixin()
-        mixin.options = Options()
-        mixin.options.enable_stats_metrics = True
-        self.assertTrue(mixin._check_stats_collection())
+        with mock.patch.dict(
+                os.environ, {
+                    "APPLICATIONINSIGHTS_STATSBEAT_DISABLED_ALL": "",
+                }):
+            self.assertTrue(mixin._check_stats_collection())
+        with mock.patch.dict(
+                os.environ, {
+                    "APPLICATIONINSIGHTS_STATSBEAT_DISABLED_ALL": "True",
+                }):
+            self.assertFalse(mixin._check_stats_collection())
         mixin._is_stats = False
         self.assertTrue(mixin._check_stats_collection())
         mixin._is_stats = True
         self.assertFalse(mixin._check_stats_collection())
-        mixin.options.enable_stats_metrics = False
+        mixin._is_stats = False
+        state._STATSBEAT_STATE["SHUTDOWN"] = False
+        self.assertTrue(mixin._check_stats_collection())
+        state._STATSBEAT_STATE["SHUTDOWN"] = True
         self.assertFalse(mixin._check_stats_collection())
+
+    def test_initial_statsbeat_success(self):
+        self.assertFalse(state._STATSBEAT_STATE["INITIAL_SUCCESS"])
+        mixin = TransportMixin()
+        mixin.options = Options()
+        mixin._is_stats = True
+        with mock.patch('requests.post') as post:
+            for code in _REACHED_INGESTION_STATUS_CODES:
+                post.return_value = MockResponse(code, 'unknown')
+                mixin._transmit([1])
+                self.assertTrue(state._STATSBEAT_STATE["INITIAL_SUCCESS"])
+                state._STATSBEAT_STATE["INITIAL_SUCCESS"] = False
+
+    def test_exception_statsbeat_shutdown_increment(self):
+        mixin = TransportMixin()
+        mixin.options = Options()
+        mixin._is_stats = True
+        state._STATSBEAT_STATE["INITIAL_SUCCESS"] = False
+        state._STATSBEAT_STATE["INITIAL_FAILURE_COUNT"] = 0
+        state._STATSBEAT_STATE["SHUTDOWN"] = False
+        with mock.patch.dict(
+                os.environ, {
+                    "APPLICATIONINSIGHTS_STATSBEAT_DISABLED_ALL": "",
+                }):
+            with mock.patch('requests.post', throw(Exception)):
+                result = mixin._transmit([1, 2, 3])
+                self.assertEqual(state._STATSBEAT_STATE["INITIAL_FAILURE_COUNT"], 1)  # noqa: E501
+                self.assertEqual(result, -1)
+
+    def test_exception_statsbeat_shutdown(self):
+        mixin = TransportMixin()
+        mixin.options = Options()
+        mixin._is_stats = True
+        state._STATSBEAT_STATE["INITIAL_SUCCESS"] = False
+        state._STATSBEAT_STATE["INITIAL_FAILURE_COUNT"] = 2
+        state._STATSBEAT_STATE["SHUTDOWN"] = False
+        with mock.patch.dict(
+                os.environ, {
+                    "APPLICATIONINSIGHTS_STATSBEAT_DISABLED_ALL": "",
+                }):
+            with mock.patch('requests.post', throw(Exception)):
+                result = mixin._transmit([1, 2, 3])
+                self.assertEqual(state._STATSBEAT_STATE["INITIAL_FAILURE_COUNT"], 3)  # noqa: E501
+                self.assertEqual(result, -2)
+
+    def test_status_code_statsbeat_shutdown_increment(self):
+        mixin = TransportMixin()
+        mixin.options = Options()
+        mixin._is_stats = True
+        state._STATSBEAT_STATE["INITIAL_SUCCESS"] = False
+        state._STATSBEAT_STATE["INITIAL_FAILURE_COUNT"] = 0
+        state._STATSBEAT_STATE["SHUTDOWN"] = False
+        with mock.patch.dict(
+                os.environ, {
+                    "APPLICATIONINSIGHTS_STATSBEAT_DISABLED_ALL": "",
+                }):
+            with mock.patch('requests.post') as post:
+                post.return_value = MockResponse(403, 'unknown')
+                mixin._transmit([1, 2, 3])
+                self.assertEqual(state._STATSBEAT_STATE["INITIAL_FAILURE_COUNT"], 1)  # noqa: E501
+                self.assertFalse(state._STATSBEAT_STATE["INITIAL_SUCCESS"])
+            with mock.patch('requests.post') as post:
+                post.return_value = MockResponse(200, 'unknown')
+                mixin._transmit([1, 2, 3])
+                self.assertEqual(state._STATSBEAT_STATE["INITIAL_FAILURE_COUNT"], 1)  # noqa: E501
+                self.assertTrue(state._STATSBEAT_STATE["INITIAL_SUCCESS"])
+
+    def test_status_code_statsbeat_shutdown(self):
+        mixin = TransportMixin()
+        mixin.options = Options()
+        mixin._is_stats = True
+        state._STATSBEAT_STATE["INITIAL_SUCCESS"] = False
+        state._STATSBEAT_STATE["INITIAL_FAILURE_COUNT"] = 2
+        state._STATSBEAT_STATE["SHUTDOWN"] = False
+        with mock.patch.dict(
+                os.environ, {
+                    "APPLICATIONINSIGHTS_STATSBEAT_DISABLED_ALL": "",
+                }):
+            with mock.patch('requests.post') as post:
+                post.return_value = MockResponse(403, 'unknown')
+                result = mixin._transmit([1, 2, 3])
+                self.assertEqual(state._STATSBEAT_STATE["INITIAL_FAILURE_COUNT"], 3)  # noqa: E501
+                self.assertFalse(state._STATSBEAT_STATE["INITIAL_SUCCESS"])
+                self.assertEqual(result, -2)
 
     def test_transmission_nothing(self):
         mixin = TransportMixin()
@@ -88,8 +193,7 @@ class TestTransportMixin(unittest.TestCase):
             self.assertIsNone(mixin.storage.get())
             self.assertEqual(len(os.listdir(mixin.storage.path)), 1)
 
-    def test_statsbeat_timeout(self):
-        _requests_map.clear()
+    def test_transmision_timeout(self):
         mixin = TransportMixin()
         mixin.options = Options()
         with mock.patch('requests.post', throw(requests.Timeout)):
@@ -99,7 +203,6 @@ class TestTransportMixin(unittest.TestCase):
             self.assertEqual(_requests_map['count'], 1)
             self.assertEqual(_requests_map['retry'], 1)
             self.assertEqual(result, mixin.options.minimum_retry_interval)
-        _requests_map.clear()
 
     def test_transmission_pre_req_exception(self):
         mixin = TransportMixin()
@@ -113,7 +216,6 @@ class TestTransportMixin(unittest.TestCase):
             self.assertEqual(len(os.listdir(mixin.storage.path)), 1)
 
     def test_statsbeat_req_exception(self):
-        _requests_map.clear()
         mixin = TransportMixin()
         mixin.options = Options()
         with mock.patch('requests.post', throw(requests.RequestException)):
@@ -123,7 +225,6 @@ class TestTransportMixin(unittest.TestCase):
             self.assertEqual(_requests_map['count'], 1)
             self.assertEqual(_requests_map['retry'], 1)
             self.assertEqual(result, mixin.options.minimum_retry_interval)
-        _requests_map.clear()
 
     def test_transmission_cred_exception(self):
         mixin = TransportMixin()
@@ -137,7 +238,6 @@ class TestTransportMixin(unittest.TestCase):
             self.assertEqual(len(os.listdir(mixin.storage.path)), 0)
 
     def test_statsbeat_cred_exception(self):
-        _requests_map.clear()
         mixin = TransportMixin()
         mixin.options = Options()
         with mock.patch('requests.post', throw(CredentialUnavailableError)):
@@ -147,7 +247,6 @@ class TestTransportMixin(unittest.TestCase):
             self.assertEqual(_requests_map['exception'], 1)
             self.assertEqual(_requests_map['count'], 1)
             self.assertEqual(result, -1)
-        _requests_map.clear()
 
     def test_transmission_client_exception(self):
         mixin = TransportMixin()
@@ -161,7 +260,6 @@ class TestTransportMixin(unittest.TestCase):
             self.assertEqual(len(os.listdir(mixin.storage.path)), 1)
 
     def test_statsbeat_client_exception(self):
-        _requests_map.clear()
         mixin = TransportMixin()
         mixin.options = Options()
         with mock.patch('requests.post', throw(ClientAuthenticationError)):
@@ -171,7 +269,6 @@ class TestTransportMixin(unittest.TestCase):
             self.assertEqual(_requests_map['retry'], 1)
             self.assertEqual(_requests_map['count'], 1)
             self.assertEqual(result, mixin.options.minimum_retry_interval)
-        _requests_map.clear()
 
     def test_transmission_pre_exception(self):
         mixin = TransportMixin()
@@ -185,7 +282,6 @@ class TestTransportMixin(unittest.TestCase):
             self.assertEqual(len(os.listdir(mixin.storage.path)), 0)
 
     def test_statsbeat_exception(self):
-        _requests_map.clear()
         mixin = TransportMixin()
         mixin.options = Options()
         with mock.patch('requests.post', throw(Exception)):
@@ -195,7 +291,6 @@ class TestTransportMixin(unittest.TestCase):
             self.assertEqual(_requests_map['exception'], 1)
             self.assertEqual(_requests_map['count'], 1)
             self.assertEqual(result, -1)
-        _requests_map.clear()
 
     @mock.patch('requests.post', return_value=mock.Mock())
     def test_transmission_lease_failure(self, requests_mock):
@@ -239,7 +334,6 @@ class TestTransportMixin(unittest.TestCase):
             self.assertEqual(len(os.listdir(mixin.storage.path)), 0)
 
     def test_statsbeat_success(self):
-        _requests_map.clear()
         mixin = TransportMixin()
         mixin.options = Options()
         with mock.patch('requests.post') as post:
@@ -250,7 +344,6 @@ class TestTransportMixin(unittest.TestCase):
             self.assertEqual(_requests_map['success'], 1)
             self.assertEqual(_requests_map['count'], 1)
             self.assertEqual(result, 0)
-        _requests_map.clear()
 
     def test_transmission_auth(self):
         mixin = TransportMixin()
@@ -300,7 +393,6 @@ class TestTransportMixin(unittest.TestCase):
             self.assertEqual(len(os.listdir(mixin.storage.path)), 1)
 
     def test_statsbeat_206(self):
-        _requests_map.clear()
         mixin = TransportMixin()
         mixin.options = Options()
         with mock.patch('requests.post') as post:
@@ -312,7 +404,6 @@ class TestTransportMixin(unittest.TestCase):
             self.assertEqual(_requests_map['count'], 1)
             self.assertEqual(_requests_map['failure'], 1)
             self.assertEqual(result, mixin.options.minimum_retry_interval)
-        _requests_map.clear()
 
     def test_transmission_206_partial_retry(self):
         mixin = TransportMixin()
@@ -342,7 +433,6 @@ class TestTransportMixin(unittest.TestCase):
             self.assertEqual(mixin.storage.get().get(), (3,))
 
     def test_statsbeat_206_partial_retry(self):
-        _requests_map.clear()
         mixin = TransportMixin()
         mixin.options = Options()
         with mock.patch('requests.post') as post:
@@ -369,7 +459,6 @@ class TestTransportMixin(unittest.TestCase):
             self.assertEqual(_requests_map['count'], 1)
             self.assertEqual(_requests_map['failure'], 1)
             self.assertEqual(result, -206)
-        _requests_map.clear()
 
     def test_transmission_206_no_retry(self):
         mixin = TransportMixin()
@@ -393,7 +482,6 @@ class TestTransportMixin(unittest.TestCase):
             self.assertEqual(len(os.listdir(mixin.storage.path)), 0)
 
     def test_statsbeat_206_no_retry(self):
-        _requests_map.clear()
         mixin = TransportMixin()
         mixin.options = Options()
         with mock.patch('requests.post') as post:
@@ -414,7 +502,6 @@ class TestTransportMixin(unittest.TestCase):
             self.assertEqual(_requests_map['count'], 1)
             self.assertEqual(_requests_map['failure'], 1)
             self.assertEqual(result, -206)
-        _requests_map.clear()
 
     def test_transmission_206_bogus(self):
         mixin = TransportMixin()
@@ -450,7 +537,6 @@ class TestTransportMixin(unittest.TestCase):
             self.assertEqual(len(os.listdir(mixin.storage.path)), 1)
 
     def test_statsbeat_429(self):
-        _requests_map.clear()
         mixin = TransportMixin()
         mixin.options = Options()
         with mock.patch('requests.post') as post:
@@ -462,7 +548,6 @@ class TestTransportMixin(unittest.TestCase):
             self.assertEqual(_requests_map['count'], 1)
             self.assertEqual(_requests_map['failure'], 1)
             self.assertEqual(result, mixin.options.minimum_retry_interval)
-        _requests_map.clear()
 
     def test_transmission_500(self):
         mixin = TransportMixin()
@@ -477,7 +562,6 @@ class TestTransportMixin(unittest.TestCase):
             self.assertEqual(len(os.listdir(mixin.storage.path)), 1)
 
     def test_statsbeat_500(self):
-        _requests_map.clear()
         mixin = TransportMixin()
         mixin.options = Options()
         with mock.patch('requests.post') as post:
@@ -489,7 +573,6 @@ class TestTransportMixin(unittest.TestCase):
             self.assertEqual(_requests_map['count'], 1)
             self.assertEqual(_requests_map['failure'], 1)
             self.assertEqual(result, mixin.options.minimum_retry_interval)
-        _requests_map.clear()
 
     def test_transmission_503(self):
         mixin = TransportMixin()
@@ -504,7 +587,6 @@ class TestTransportMixin(unittest.TestCase):
             self.assertEqual(len(os.listdir(mixin.storage.path)), 1)
 
     def test_statsbeat_503(self):
-        _requests_map.clear()
         mixin = TransportMixin()
         mixin.options = Options()
         with mock.patch('requests.post') as post:
@@ -516,7 +598,6 @@ class TestTransportMixin(unittest.TestCase):
             self.assertEqual(_requests_map['count'], 1)
             self.assertEqual(_requests_map['failure'], 1)
             self.assertEqual(result, mixin.options.minimum_retry_interval)
-        _requests_map.clear()
 
     def test_transmission_401(self):
         mixin = TransportMixin()
@@ -530,7 +611,6 @@ class TestTransportMixin(unittest.TestCase):
             self.assertEqual(len(os.listdir(mixin.storage.path)), 1)
 
     def test_statsbeat_401(self):
-        _requests_map.clear()
         mixin = TransportMixin()
         mixin.options = Options()
         with mock.patch('requests.post') as post:
@@ -542,7 +622,6 @@ class TestTransportMixin(unittest.TestCase):
             self.assertEqual(_requests_map['count'], 1)
             self.assertEqual(_requests_map['failure'], 1)
             self.assertEqual(result, mixin.options.minimum_retry_interval)
-        _requests_map.clear()
 
     def test_transmission_403(self):
         mixin = TransportMixin()
@@ -556,7 +635,6 @@ class TestTransportMixin(unittest.TestCase):
             self.assertEqual(len(os.listdir(mixin.storage.path)), 1)
 
     def test_statsbeat_403(self):
-        _requests_map.clear()
         mixin = TransportMixin()
         mixin.options = Options()
         with mock.patch('requests.post') as post:
@@ -568,7 +646,6 @@ class TestTransportMixin(unittest.TestCase):
             self.assertEqual(_requests_map['count'], 1)
             self.assertEqual(_requests_map['failure'], 1)
             self.assertEqual(result, mixin.options.minimum_retry_interval)
-        _requests_map.clear()
 
     def test_transmission_307(self):
         mixin = TransportMixin()
@@ -598,7 +675,6 @@ class TestTransportMixin(unittest.TestCase):
         self.assertEqual(mixin.options.endpoint, "https://example.com")
 
     def test_statsbeat_307(self):
-        _requests_map.clear()
         mixin = TransportMixin()
         mixin.options = Options()
         mixin._consecutive_redirects = 0
@@ -612,7 +688,6 @@ class TestTransportMixin(unittest.TestCase):
             self.assertEqual(_requests_map['count'], 10)
             self.assertEqual(_requests_map['failure'], 10)
             self.assertEqual(result, -307)
-        _requests_map.clear()
 
     def test_transmission_439(self):
         mixin = TransportMixin()
@@ -626,7 +701,6 @@ class TestTransportMixin(unittest.TestCase):
             self.assertEqual(len(os.listdir(mixin.storage.path)), 0)
 
     def test_statsbeat_439(self):
-        _requests_map.clear()
         mixin = TransportMixin()
         mixin.options = Options()
         with mock.patch('requests.post') as post:
@@ -637,4 +711,3 @@ class TestTransportMixin(unittest.TestCase):
             self.assertEqual(_requests_map['throttle'], 1)
             self.assertEqual(_requests_map['count'], 1)
             self.assertEqual(result, -439)
-        _requests_map.clear()